feat: add ability to commit suggested message
Adds the ability to commit the suggested commit message. If there is
only one suggestion, the program will confirm with the user whether they
want to commit with that message. If there are multiple suggestions, the
user will be prompted to choose which message to commit with.

This commit also adds the `commit` function that handles the `git
commit` command.