feat(cli): add temperature and frequency penalty options
This commit adds the `-t` and `-f` options for setting the temperature and frequency penalty values to pass into the OpenAI API request. The `t` value must be a float between 0.0 and 2.0, while `f` must be between 0.0 and 2.0. Both values have default values in case they are not passed by the user.