chore(cli): refactor options to use Option type
Refactor the Options struct in cli.rs to use the Option type for all fields. This allows for more flexibility in the command line arguments that can be passed in. Also, update the help message to include information on the new -f flag, which is used to adjust the frequency of the AI's responses.