feat(config): add ability to customize default options
This commit adds a configuration file to the program that allows users to customize the default options for temperature, frequency penalty, number of choices, and system message prompt. The configuration file is stored in the home directory as `.turbocommit.yaml` and can be removed to reset to the default system message. The configuration can be accessed by running the program and following the instructions in the terminal.