feat: Allow editing of commit message before committing
This change adds the ability to edit the commit message and commit only if the user is satisfied with the change. The user is prompted to select a commit message from a list of choices, and they can then choose to edit the message, commit with the selected message or abort the operation.

fix #13